feat(travis): upload docs and client to S3
This change uses Travis CI's artifact feature to store build information
in S3. Specifically we'd like a client and a doc build so that we can test
and view changes for a specific build with these artifacts, as the rest of
the build artifacts are Docker images, uploaded to Quay.io and DockerHub.