ref(models): add domains to k8s data for router to use
Domain information is now saved to routerConfig annotations in addition to still being in etcd. This solves part of #50
Also adds new flatten and dict_merge to both scheduler and models

The code has a little oddity in it where subdomains for app (on the platform domain) are added during scale event from 0 to 1. That will be address in #88
It is possible to remove the apps subdomain but during any scaling event it gets added back to the Domain model. This allows the router to treat all domains the same way