feat(travis): deploy client builds to bintray
Travis artifacts is great for creating builds and uploading to S3,
but one thing it misses on is the ability to grab the "latest" build
(as seen in the README). In order for a user to grab the latest
client build, he or she must know the latest Travis build that went
green and download the client using that URL.

Seth Goings fronted the work with releasing the helm client on
bintray, which so far has been a great success. This PR moves us
over to bintray for that one-liner `(curl https://... | sh)` special
sauce.