feat(healthcheck): use k8s liveness and readiness probes for app healthchecks
Use k8s liveness and readiness probes (HTTP GET only for now) to replace the defunct publisher healthcheck system. For now those two are conflated into one to support the current documented deis healthcheck workflow and an assumption of port 80 is in the code, as well as HTTP (no ssl).

All healthcheck related tests were axed as they are no longer applicable and in fact not possible to test until there is a more robust k8s testing strategy for workflow.

http://kubernetes.io/v1.1/docs/user-guide/pod-states.html#container-probes has more information on what probes can do and what's available

Fixes #76