chore(quickstart): refactor quickstart
Skeleton refactor to test new provider-specific flows.

One of the struggles I've had with the quickstart is that it has been
caught between among a task-oriented doc, platform reference and
provider agnostic. The end-result is something that isn't compelling and
ends up being particularly confusing for new visitors to the project.

This (now squashed) commit represents a deep refactor of the quickstart
documentation.

Orient the quickstart around a three step, goal-oriented process:

1. Tools
2. Infra + Workflow Install
3. Deploy an example app

This also pulls in specific configurations of Kubernetes for evaluation
and dev/test (kube-up.sh and GKE). We want tire-kickers to get a win and
we need to offer as much assistance through the process as possible.

Pre-install helm and deis CLIs up front.

Couching the install in provider-specific workflows gives the
documentation much needed context. For example, nuances about load
balancer configurations (or lack there-of) can be dealt with in-context.
This will simplify the language as we can back away from routing folks
to more detailed explanations or landing a barrage of caveats.

I also slimmed down the "deploy your first app" to the absolute basics.
Using the `deis/example-go` app lets folks get a win without having to
deal with cloning repos, etc. That stuff can go into reference
documentation for the three build types.