Add instructions to downgrade GKE node versions
After spending too much time on installing Workflow, @kmala pointed out in Slack that I was probably running on k8s 1.3.x. Which is the default. I checked in the Console and there is no other way than to create the new cluster then downgrade it.

Hope with this (temporary) fix in the docs, other people will not make the same mistake.